--- webui/react/src/hooks/useModal/Experiment/useModalExperimentMove.module.scss
@@ -1,11 +1,0 @@
-.base {
-  color: var(--theme-float-on);
-}
-.optionDisabled {
-  display: flex;
-  justify-content: space-between;
-
-  span {
-    color: var(--theme-surface-on-weak);
-  }
-}
